% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthonyverse-package.R
\docType{package}
\name{anthonyverse-package}
\alias{anthonyverse}
\alias{anthonyverse-package}
\title{anthonyverse: Various Helper Functions for Data Analysis}
\description{
A collection of functions that I use for data analysis. This package is a work in progress and will be updated frequently.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anthonyokc/anthonyverse/}
  \item Report bugs at \url{https://github.com/anthonyokc/anthonyverse/issues}
}

}
\author{
\strong{Maintainer}: anthonyokc \email{anthonyokc@example.com}

}
\keyword{internal}
