% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.R
\name{comma}
\alias{comma}
\title{FUNCTION_TITLE}
\usage{
comma(x, digits = 0, abs = FALSE, round_to_zero = FALSE, drop0trailing = FALSE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{digits}{PARAM_DESCRIPTION, Default: 0}

\item{abs}{PARAM_DESCRIPTION, Default: FALSE}

\item{round_to_zero}{PARAM_DESCRIPTION, Default: FALSE}

\item{drop0trailing}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[janitor]{round_half_up}}
}
\author{
Anthony Flores
}
