% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_eviction_labels.R
\name{covid_eviction_labels}
\alias{covid_eviction_labels}
\title{FUNCTION_TITLE}
\usage{
covid_eviction_labels(label_data, ..., .labels = FALSE)
}
\arguments{
\item{label_data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{.labels}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_label}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{annotate}}, \code{\link[ggplot2]{geom_raster}}, \code{\link[ggplot2]{scale_manual}}, \code{\link[ggplot2]{guide_legend}}, \code{\link[ggplot2]{geom_abline}}
\code{\link[lubridate]{ymd}}
}
\author{
Anthony Flores
}
