% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_theme_ojo.R
\name{hc_theme_ojo}
\alias{hc_theme_ojo}
\title{FUNCTION_TITLE}
\usage{
hc_theme_ojo()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[highcharter]{hc_theme_merge}}, \code{\link[highcharter]{hc_theme_smpl}}, \code{\link[highcharter]{hc_theme}}
\code{\link[ojodb]{ojo_pal}}
}
\author{
Anthony Flores
}
