% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart_label_formatter.R
\name{highchart_label_formatter}
\alias{highchart_label_formatter}
\title{FUNCTION_TITLE}
\usage{
highchart_label_formatter(
  variable_name,
  prefix = "",
  suffix = "",
  value_identifier = "this.value",
  digits = 0,
  dataClass = FALSE
)
}
\arguments{
\item{variable_name}{PARAM_DESCRIPTION}

\item{prefix}{PARAM_DESCRIPTION, Default: ''}

\item{suffix}{PARAM_DESCRIPTION, Default: ''}

\item{value_identifier}{PARAM_DESCRIPTION, Default: 'this.value'}

\item{digits}{PARAM_DESCRIPTION, Default: 0}

\item{dataClass}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_detect}}, \code{\link[stringr]{str_replace}}
\code{\link[htmlwidgets]{JS}}
}
\author{
Anthony Flores
}
