% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_add_geoid.R
\name{ojo_add_geoid}
\alias{ojo_add_geoid}
\title{FUNCTION_TITLE}
\usage{
ojo_add_geoid(data, geography = c("county", "state"), year = 2019)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{geography}{PARAM_DESCRIPTION, Default: c("county", "state")}

\item{year}{PARAM_DESCRIPTION, Default: 2019}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{is_installed}}
\code{\link[tidycensus]{get_estimates}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate-joins}}
\code{\link[stringr]{str_remove}}, \code{\link[stringr]{case}}, \code{\link[stringr]{str_replace}}
}
\author{
Anthony Flores
}
