% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_add_geometries.R
\name{ojo_add_geometries}
\alias{ojo_add_geometries}
\title{FUNCTION_TITLE}
\usage{
ojo_add_geometries(data, geography = c("county", "state"), ...)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{geography}{PARAM_DESCRIPTION, Default: c("county", "state")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tigris]{counties}}, \code{\link[tigris]{states}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{cross_join}}
}
\author{
Anthony Flores
}
