% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_add_month_and_year.R
\name{ojo_add_month_and_year}
\alias{ojo_add_month_and_year}
\title{FUNCTION_TITLE}
\usage{
ojo_add_month_and_year(data, ...)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{defusing-advanced}}, \code{\link[rlang]{quo_label}}, \code{\link[rlang]{sym}}
\code{\link[purrr]{map}}
\code{\link[dplyr]{mutate}}
\code{\link[lubridate]{round_date}}
}
\author{
Anthony Flores
}
