% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_add_population.R
\name{ojo_add_population}
\alias{ojo_add_population}
\title{FUNCTION_TITLE}
\usage{
ojo_add_population(
  data,
  geography = c("both", "county", "state"),
  year_column,
  individual_year = NA,
  ...
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{geography}{PARAM_DESCRIPTION, Default: c("both", "county", "state")}

\item{year_column}{PARAM_DESCRIPTION}

\item{individual_year}{PARAM_DESCRIPTION, Default: NA}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{nth}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{join_by}}
\code{\link[lubridate]{year}}, \code{\link[lubridate]{ymd}}
\code{\link[tidycensus]{get_estimates}}
\code{\link[tidyr]{pivot_wider}}
}
\author{
Anthony Flores
}
