% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_aggregate.R
\name{ojo_aggregate}
\alias{ojo_aggregate}
\title{FUNCTION_TITLE}
\usage{
ojo_aggregate(
  data,
  ...,
  group_by,
  geography = c("county", "state"),
  rate = 10^3,
  rate_suffix = "1k"
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{group_by}{PARAM_DESCRIPTION}

\item{geography}{PARAM_DESCRIPTION, Default: c("county", "state")}

\item{rate}{PARAM_DESCRIPTION, Default: 10^3}

\item{rate_suffix}{PARAM_DESCRIPTION, Default: '1k'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}}, \code{\link[rlang]{quosure-tools}}, \code{\link[rlang]{as_label}}
\code{\link[purrr]{map}}
\code{\link[dplyr]{if_else}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}
}
\author{
Anthony Flores
}
