% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_district_and_date_filters.R
\name{ojo_district_and_date_filters}
\alias{ojo_district_and_date_filters}
\title{FUNCTION_TITLE}
\usage{
ojo_district_and_date_filters(
  data,
  districts = "ALL",
  date_start = NA,
  date_end = NA,
  district_variable = "district",
  date_variable = "date_filed"
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{districts}{PARAM_DESCRIPTION, Default: 'ALL'}

\item{date_start}{PARAM_DESCRIPTION, Default: NA}

\item{date_end}{PARAM_DESCRIPTION, Default: NA}

\item{district_variable}{PARAM_DESCRIPTION, Default: 'district'}

\item{date_variable}{PARAM_DESCRIPTION, Default: 'date_filed'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
\author{
Anthony Flores
}
