% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_optional_filter.R
\name{ojo_optional_filter}
\alias{ojo_optional_filter}
\title{FUNCTION_TITLE}
\usage{
ojo_optional_filter(data, variable, variable_values = NA)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION}

\item{variable_values}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
\author{
Anthony Flores
}
