% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{FUNCTION_TITLE}
\usage{
percent(
  x,
  digits = 0,
  abs = FALSE,
  drop0trailing = FALSE,
  multiply_by_100 = TRUE,
  ...
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{digits}{PARAM_DESCRIPTION, Default: 0}

\item{abs}{PARAM_DESCRIPTION, Default: FALSE}

\item{drop0trailing}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[janitor]{round_half_up}}
}
\author{
Anthony Flores
}
