% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_variable_names.R
\name{pretty_variable_names}
\alias{pretty_variable_names}
\title{FUNCTION_TITLE}
\usage{
pretty_variable_names(variable_names)
}
\arguments{
\item{variable_names}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[snakecase]{caseconverter}}
\code{\link[stringr]{str_replace}}
}
\author{
Anthony Flores
}
