% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_percent_missing.R
\name{report_percent_missing}
\alias{report_percent_missing}
\title{FUNCTION_TITLE}
\usage{
report_percent_missing(data, column)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{column}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{nth}}, \code{\link[dplyr]{pull}}
}
\author{
Anthony Flores
}
