% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tably.R
\name{tably}
\alias{tably}
\title{FUNCTION_TITLE}
\usage{
tably(data, ..., print_n = NA, top_n = NA, sort = TRUE)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{print_n}{PARAM_DESCRIPTION, Default: NA}

\item{top_n}{PARAM_DESCRIPTION, Default: NA}

\item{sort}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}}, \code{\link[rlang]{sym}}
\code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{summarise}}
}
\author{
Anthony Flores
}
